Refactor how the schema is set up, to improve code generation result
Split up the OpenAPI definition: Only the basic information about the
endpoint remain in openapi.yml. All the schemas are now in a separate
file schema.yml that gets referenced from openapi.yml.

All the sub-schemas are now part of the JSON-Schema `definitions`. This
way, we don't need to de-reference the entire OpenAPI specification file
to generate a schema-file that can be used as input to other files.

Therefore, we can just use our schema-file with the `definitions` as
input for the code generation. As a result, the code generator now
correctly picks up on sub-schemas that are re-used in multiple types.
For example, we now have a definition `Person` that gets used in both
the `creator` and `performer` fields, and the code generator knows to
just generate one Elm-Type for the `Person` type.