Re-structure schema.yml to fix auto-generated code
Unfortunately, the automatic code generation has some issues, especially
when two objects use the same value for the @type-field. E.g. both
`Event` and `AdditionalOffering` are of type "Event", but this leads to
conflicting auto-generated code.
As a workaround, the schema was re-structured slightly to avoid this and
similar issues.