Fix grid layout
The fixed-grid apparently does not play well with cells spanning
multiple columns. When the second column had a lot of content (i.e.
there  was a long list of people in the team) it would grow so wide as
to almost completely minimize the first column. I think it's related to
the fixed-grid breakpoints being based on the container width vs. the
cell breakpoints being based on the screen/window width.

Anyway, using the smart grid instead does not have the same issue and
works well enough.