Support multi-language strings for some fields on Production
This commit adds a new InternationalizedString schema, that allows for
some fields to be supplied with a multi-lingual text object. Since we
don't want to break validity of existing implementations of the API,
this is an optional feature. I.e. all of these fields also still accept
a simple string.
The rendering code was also updated to support fields that contain,
potentially, multiple translations of the same text.